Move SingleTableInventoryCalculator and related classes
